+++ apps/web/components/scanner/index.tsx
@@ -1,0 +1,4 @@
+export * from './footer';
+export * from './pointer';
+export * from './stage';
+export * from './toolbar';
