+++ packages/types/src/api/index.ts
@@ -1,2 +1,3 @@
 export * from './express.types';
 export * from './invite.types';
+export * from './upload.types';
