+++ packages/types/src/api/upload.types.ts
@@ -1,0 +1,12 @@
+import { coordinates } from '..';
+
+export interface IUploadFile {
+  url: string;
+  name: string;
+  border: {
+    topLeft: coordinates;
+    topRight: coordinates;
+    bottomLeft: coordinates;
+    bottomRight: coordinates;
+  };
+}
