+++ packages/types/src/web/scanner.types.ts
@@ -1,0 +1,4 @@
+export type coordinates = {
+  x: number;
+  y: number;
+};
